+++ c10/core/TensorOptions.h
@@ -5,7 +5,7 @@
 #include <c10/core/Layout.h>
 #include <c10/core/ScalarType.h>
 #include <c10/core/Device.h>
-#include <c10/core/TensorTypeSet.h>
+#include <c10/core/DispatchKeySet.h>
 
 #include <c10/util/Optional.h>
 #include <c10/util/C++17.h>
